more tests and preparing for take refactor
