+++ polars/polars-core/src/chunked_array/ops/chunkops.rs
@@ -128,3 +128,18 @@ where
         }
     }
 }
+
+#[cfg(test)]
+mod test {
+    use crate::prelude::*;
+
+    #[test]
+    fn test_categorical_map_after_rechunk() {
+        let s = Series::new("", &["foo", "bar", "spam"]);
+        let mut a = s.cast::<CategoricalType>().unwrap();
+
+        a.append(&a.slice(0, 2).unwrap());
+        a.rechunk();
+        assert!(a.categorical().unwrap().categorical_map.is_some());
+    }
+}
